[front] refactor: cleaner code for topBar
